chore(type): rename files for state typing
