+++ types/store.d.ts
@@ -1,6 +1,6 @@
-import { NLife } from './life'
-import { NStatus } from './status'
-import { NVenues } from './venues'
+import { NLife } from './lifeState'
+import { NStatus } from './statusState'
+import { NVenues } from './venuesState'
 
 export declare namespace NStore {
   interface IState {
